Update naming convention on all STLs
The naming convention for STL files has been updated to include the version number of the part, as well as the date it was released.

Designs receiving fixes, enhancements, or printability improvements, will now maintain the same SKU and increment the version.

In cases where a design undergoes a significant redesign, a new SKU will be assigned.

In all cases, previous versions will be relocated to the "Deprecated" subfolder.